requirements.txt
flask
requests
scikit-learn
python-docx
PyPDF2